+++ src/styles/PostList.module.css
@@ -23,7 +23,6 @@
 
 .table th {
     background-color: white;
-    /* color: white; */
     font-weight: bold;
     text-transform: uppercase;
     border-bottom: 2px solid #ddd; /* 헤더와 본문 사이의 두꺼운 하단 경계선 */
@@ -50,35 +49,10 @@
 .table tr:hover {
     background-color: #f1f1f1;
 }
-.pagination {
-    display: flex;
-    justify-content: center;
-    margin-top: 20px;
-}
 
-.pagination button {
-    background-color: #864e5d;
-    color: white;
-    border: none;
-    padding: 10px 20px;
-    margin: 0 5px;
-    cursor: pointer;
-    border-radius: 4px;
-    font-size: 16px;
-}
-
-.pagination button:hover {
-    background-color: #673d48;
-}
-
-.pagination button.active {
-    background-color: #333;
-    color: white;
-}
 .clickableRow {
     cursor: pointer; /* 마우스 커서를 링크 클릭할 때처럼 바꾸기 위한 스타일 */
 }
-
 
 /* 카테고리 열 너비 줄이기 */
 .table th:nth-child(1),
